+++ src/main/java/com/brano/ecomonitor/repository/CriticalOrganRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CriticalOrganRepository extends JpaRepository<CriticalOrgan, Long> {
+public interface CriticalOrganRepository extends JpaRepository<CriticalOrgan, Integer> {
 }
