+++ src/main/java/com/brano/ecomonitor/repository/PollutantRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PollutantRepository extends JpaRepository<Pollutant, Long> {
+public interface PollutantRepository extends JpaRepository<Pollutant, Integer> {
 }
